requirements.txt
vectorbt
numpy
alpaca_trade_api
alpaca_py
kaleido
pandas_ta
loguru